backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.30.1
nltk==3.8.1
# Optional: for SVG to PNG conversion
# cairosvg==2.7.0
